I added little hack to kern the MenuFont in PutStringFont function..
it was a bit loose looking on the screen, and if I would have
adjusted the spacing only in the font .png file, we would have
lost the glow.

anyways, I tried to make a real "kerning" function to BFont,
(SetFontKerning)
but didn't manage to make it function properly :)
so I removed everything and did a simple hack.

I didn't want to change the code too much, because then it would
have required a lot of adjusting the code elsewhere whenever
bfont is used.. so...


.b