ok. here are my changes for this evening. I started turning the city more city like.. it has some sort of roads, and sidewalks in it..  also, I decorated some of the houses a little better, so that they actually look like someone lives in them.
I also removed toilets from peoples houses.. who on earth would have
that in his living room? ;)

I redid the whole wall that goes around the city. it had some bugs in it,
and there was strange double walls.. I tried to get rid of those.

The level editor gives me headaches, seems to hog memory. after some time of using,
it is so slow that I need to quickly save and kill the whole program.

anyways, I decided to submit this now, even if it's not ready, but, just to be
on safe side. I'm not sure how much I have time for this tomorrow.

.b