I do this on my own, without consultation with anyone. If anyone is to get into trouble for this, it is just me, and no one else.
We have 8 music files in the game. Six made by Arvid.
Out of the other two, one is not in use anymore. (Bleostrada.ogg)
This makes Arvid the maker of 6/7ths of the game music.
This is about 85% of the game music.
He is not going to get 0% credit for his work, as long as I am a team member.
This is not justice. This is not logical. This is not fair.

If we exclude people who have done less than 85% of work on their branch, the
credits are going to be damn empty.

If this gets reverted, please also delete lines 3062 and 3063 from the same file.

That is all.